+++ discord/voice/client.py
@@ -30,7 +30,12 @@ import datetime
 import logging
 import struct
 import warnings
-from typing import TYPE_CHECKING, Any, Literal, overload
+from collections.abc import Callable, Coroutine
+from typing import TYPE_CHECKING, Any, Literal, cast, overload
+
+import nacl.secret
+import nacl.utils
+from typing_extensions import Never, Unpack
 
 from discord import opus
 from discord.enums import SpeakingState, try_enum
@@ -40,19 +45,20 @@ from discord.sinks.core import Sink
 from discord.sinks.errors import RecordingException
 from discord.utils import MISSING
 
-from ..utils import get_missing_voice_dependencies
 from ._types import VoiceProtocol
 from .enums import OpCodes
 from .receive import AudioReader
 from .state import VoiceConnectionState
-from .utils.dependencies import HAS_DAVEY, HAS_NACL
+from .utils.dependencies import has_nacl
 
-if HAS_NACL:
+if has_nacl:
     import nacl.secret
     import nacl.utils
 
 if TYPE_CHECKING:
-    from typing_extensions import ParamSpec
+    from typing import TypeVar
+
+    from typing_extensions import ParamSpec, TypeVarTuple
 
     from discord import abc
     from discord.client import Client
@@ -71,13 +77,15 @@ if TYPE_CHECKING:
     from .receive.reader import AfterCallback
 
     P = ParamSpec("P")
+    T = TypeVar("T")
+    Ts = TypeVarTuple("Ts")
 
 _log = logging.getLogger(__name__)
 
 __all__ = ("VoiceClient",)
 
 
-class VoiceClient(VoiceProtocol):
+class VoiceClient(VoiceProtocol["Client"]):
     """Represents a Discord voice connection.
 
     You do not create these, you typically get them from e.g.
@@ -107,15 +115,6 @@ class VoiceClient(VoiceProtocol):
         client: Client,
         channel: abc.Connectable,
     ) -> None:
-        missing = get_missing_voice_dependencies()
-        if missing:
-            deps = ", ".join(missing)
-            raise RuntimeError(
-                f"{deps} {'library is' if len(missing) == 1 else 'libraries are'} needed "
-                "in order to use voice related features, "
-                'you can run "pip install py-cord[voice]" to install all voice-related '
-                "dependencies."
-            )
 
         super().__init__(client, channel)
         state = client._connection
@@ -136,8 +135,10 @@ class VoiceClient(VoiceProtocol):
 
         self._ssrc_to_id: dict[int, int] = {}
         self._id_to_ssrc: dict[int, int] = {}
-        self._event_listeners: dict[str, list] = {}
-        self._reader: AudioReader = MISSING
+        self._event_listeners: dict[
+            str, list[Callable[..., Coroutine[Any, Any, Any]]]
+        ] = {}
+        self._reader: AudioReader[Any] = MISSING
 
     @staticmethod
     def _set_future_result_if_pending(
@@ -156,7 +157,7 @@ class VoiceClient(VoiceProtocol):
     @property
     def guild(self) -> Guild:
         """Returns the guild the channel we're connecting to is bound to."""
-        channel: VocalGuildChannel = self.channel
+        channel = cast("VocalGuildChannel", self.channel)
         return channel.guild
 
     @property
@@ -259,7 +260,11 @@ class VoiceClient(VoiceProtocol):
         # maybe handle video and such things?
 
     async def _run_event(
-        self, coro, event_name: str, *args: Any, **kwargs: Any
+        self,
+        coro: Callable[..., Coroutine[P, None]],
+        event_name: str,
+        *args: P.args,
+        **kwargs: P.kwargs,
     ) -> None:
         try:
             await coro(*args, **kwargs)
@@ -269,8 +274,12 @@ class VoiceClient(VoiceProtocol):
             _log.exception("Error calling %s", event_name)
 
     def _schedule_event(
-        self, coro, event_name: str, *args: Any, **kwargs: Any
-    ) -> asyncio.Task:
+        self,
+        coro: Callable[..., Coroutine[Any, Any, T]],
+        event_name: str,
+        *args: Any,
+        **kwargs: Any,
+    ) -> asyncio.Task[T]:
         wrapped = self._run_event(coro, event_name, *args, **kwargs)
         return self.client.loop.create_task(
             wrapped, name=f"voice-receiver-event-dispatch: {event_name}"
@@ -433,6 +442,7 @@ class VoiceClient(VoiceProtocol):
         return encrypt_packet(header, packet)
 
     # encryption methods
+    # nacl is guaranteed to be available here because __init__ raises if missing
 
     def _encrypt_xsalsa20_poly1305(self, header: bytes, data: Any) -> bytes:
         # deprecated
@@ -569,8 +579,10 @@ class VoiceClient(VoiceProtocol):
             raise ClientException("Not connected to voice")
         if self.is_playing():
             raise ClientException("Already playing audio")
-        if not isinstance(source, AudioSource):
-            raise TypeError(
+        if not isinstance(
+            source, AudioSource
+        ):  # pyright: ignore[reportUnnecessaryIsInstance]
+            raise TypeError(  # pyright: ignore[reportUnreachable]
                 f"Source must be an AudioSource, not {source.__class__.__name__}",
             )
         if not self.encoder and not source.is_opus():
@@ -636,8 +648,12 @@ class VoiceClient(VoiceProtocol):
 
     @source.setter
     def source(self, value: AudioSource) -> None:
-        if not isinstance(value, AudioSource):
-            raise TypeError(f"expected AudioSource, not {value.__class__.__name__}")
+        if not isinstance(
+            value, AudioSource
+        ):  # pyright: ignore[reportUnnecessaryIsInstance]
+            raise TypeError(
+                f"expected AudioSource, not {value.__class__.__name__}"
+            )  # pyright: ignore[reportUnreachable]
 
         if self._player is None:
             raise ValueError("the client is not playing anything")
@@ -688,36 +704,44 @@ class VoiceClient(VoiceProtocol):
             return datetime.timedelta(milliseconds=self._player.played_frames() * 20)
         return datetime.timedelta()
 
+    @overload
     def start_recording(
         self,
         sink: Sink,
-        callback: AfterCallback | None = None,
-        *args: Any,
+        callback: None = None,
+        *args: None,
+        sync_start: bool = MISSING,
+    ) -> None: ...
+
+    @overload
+    def start_recording(
+        self,
+        sink: Sink,
+        callback: AfterCallback[Unpack[Ts]],
+        *args: tuple[Unpack[Ts]],
+        sync_start: bool = MISSING,
+    ) -> None: ...
+
+    def start_recording(
+        self,
+        sink: Sink,
+        callback: AfterCallback[Unpack[Ts]] | None = None,
+        *args: tuple[Unpack[Ts]] | None,
         sync_start: bool = MISSING,
     ) -> None:
         r"""Start recording the audio from the current connected channel to the provided sink.
 
         .. versionadded:: 2.0
 
-        .. warning::
-
-            Recording may not work as expected due to the new DAVE (End-to-End Encryption) for voice calls.
-
         Parameters
         ----------
         sink: :class:`~.Sink`
             A Sink in which all audio packets will be processed in.
-        callback: Callable[[:class:`Exception` | None], Any]
-            A function which is called after the bot has stopped recording. This must take exactly one positonal(-only)
-            parameter, ``exception``, which is the exception that was raised during the recording of the Sink.
-
-            .. versionchanged:: 2.7
-                This parameter is now optional, and must take exactly one parameter, ``exception``.
+        callback:
+            A (potentially async) function which is called after the bot has stopped recording.
+            The function must take the sink as its first parameter.
         \*args:
-            The arguments to pass to the callback coroutine.
-
-            .. deprecated:: 2.7
-                Passing custom arguments to the callback is now deprecated and ignored.
+            Any additional positional arguments to pass to the callback coroutine.
         sync_start: :class:`bool`
             If ``True``, the recordings of subsequent users will start with silence.
             This is useful for recording audio just as it was heard.
@@ -732,17 +756,12 @@ class VoiceClient(VoiceProtocol):
         TypeError
             You did not provide a Sink object.
         """
-        warnings.warn(
-            "Voice reception is currently broken due to Discord's DAVE (End-to-End Encryption) protocol. "
-            + "Follow development progress at https://github.com/Pycord-Development/pycord/issues/3139",
-            RuntimeWarning,
-            stacklevel=2,
-        )
-        # TODO: remove warning in voice-recv fix PR
         if not self.is_connected():
             raise RecordingException("not connected to a voice channel")
-        if not isinstance(sink, Sink):
-            raise TypeError(f"expected a Sink object, got {sink.__class__.__name__}")
+        if not isinstance(sink, Sink):  # pyright: ignore[reportUnnecessaryIsInstance]
+            raise TypeError(
+                f"expected a Sink object, got {sink.__class__.__name__}"
+            )  # pyright: ignore[reportUnreachable]
 
         if self.is_recording():
             raise ClientException("Already recording audio")
@@ -756,7 +775,9 @@ class VoiceClient(VoiceProtocol):
                 "'sync_start' parameter is deprecated since 2.7 and will be removed in 3.0"
             )
 
-        self._reader = AudioReader(sink, self, after=callback, start=True)
+        self._reader = AudioReader(
+            sink, self, after=callback, args=args, start=True
+        )  # pyright: ignore[reportCallIssue, reportAttributeAccessIssue, reportArgumentType]
 
     start_listening = start_recording
 
@@ -770,12 +791,6 @@ class VoiceClient(VoiceProtocol):
         RecordingException
             You are not recording.
         """
-        warnings.warn(
-            "Voice reception is currently broken due to Discord's DAVE (End-to-End Encryption) protocol. "
-            + "Follow development progress at https://github.com/Pycord-Development/pycord/issues/3139",
-            RuntimeWarning,
-            stacklevel=2,
-        )
         if self._reader is not MISSING:
             self._reader.stop()
             self._reader = MISSING
@@ -797,12 +812,6 @@ class VoiceClient(VoiceProtocol):
 
         .. versionadded:: 2.7
         """
-        warnings.warn(
-            "Voice reception is currently broken due to Discord's DAVE (End-to-End Encryption) protocol. "
-            + "Follow development progress at https://github.com/Pycord-Development/pycord/issues/3139",
-            RuntimeWarning,
-            stacklevel=2,
-        )
         ssrc = self._id_to_ssrc.get(member.id)
         if ssrc is None:
             return None